+++ react-frontend/src/components/TaskInfo.js
